Introduce isItFavoritesTable variable to differentiate table types
Introduced a new variable, isItFavoritesTable, within the characterTable to distinguish between different table types. This variable is assigned during the creation of the table and is used to control certain conditions, enabling more tailored functionalities based on whether the table is displaying favorites or a general character list. This addition improves logic handling and enhances the flexibility of table view behavior.