Added functionality to handle character details opening via viewModel.
Added viewModel.onCharacterDetailsOpened closure to handle character details opening. When triggered, it calls the toggleDetailsView function, which presents CHDetailsViewController with the appropriate CHDetailsViewModel initialized with the selected character. The presented view controller is embedded in a UINavigationController with modal presentation style set to .pageSheet. Additionally, a dismissSheet closure is implemented in CHDetailsViewController to refresh the view when dismissed. This enhancement improves the organization and flow of character details presentation in the application.