Added empty state texts and titles to provide user feedback when no favorite characters are added, and implemented character details toggling functionality.
Empty state texts and titles, named emptyStateText and emptyStateTitle, were added to inform users when no favorite characters are added. These views are initially hidden and become visible when the list of characters is empty. Additionally, implemented functionality to toggle character details view was added. This functionality creates and presents a CHDetailsViewController with the selected character's details. The presented view controller is embedded in a UINavigationController with modal presentation style set to .pageSheet. A dismissSheet closure is implemented in CHDetailsViewController to refresh the view when dismissed. These changes enhance the user experience by providing better feedback and enabling character details viewing.